third_party/WebKit/LayoutTests/http/tests/websocket/iframe-sandbox.html

<!DOCTYPE html>
<script src="/js-test-resources/js-test.js"></script>
<script>
description("WebSocket test that Origin is 'null' in a sandboxed iframe.");

window.jsTestIsAsync = true;

window.addEventListener("message", receiveMessage, false);

var wsOrigin;

function receiveMessage(event)
{
    wsOrigin = event.data;
    shouldBeEqualToString("wsOrigin", "null");
    finishJSTest();
}
</script>
<iframe sandbox="allow-scripts" src="resources/iframe-sandbox-iframe.html">
</iframe>
